refactored routes into separate routers
